Aligning Zassenhaus impl with Trotter
